3 only
The new graphic looks good and so do the added copy for five. There is
just one thing still bothering me. I still think we need and answer to
my questions at the end of 3.2. If it was ignored because you don’t—at
this point—know the answer, then just flag it somehow. What does the
bug rate under 1 mean? And how did the analysis identify B as the buggy
set.